# apps/api/internal/handlers/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PulseWatch API",
    "version": "1.0.0",
    "description": "REST API for PulseWatch uptime monitoring"
  },
  "servers": [{ "url": "/api/v1" }],
  "paths": {
...
    "/orgs/{orgId}/monitors/{id}/stats": { "get": { "summary": "Monitor statistics and trend", "security": [{ "bearerAuth": [] }] } },
...
    "/orgs/{orgId}/status-pages": { "get": { "summary": "List status pages", "security": [{ "bearerAuth": [] }] }, "post": { "summary": "Create status page", "security": [{ "bearerAuth": [] }] } },
    "/orgs/{orgId}/status-pages/{pageId}": { "get": { "summary": "Get status page", "security": [{ "bearerAuth": [] }] }, "patch": { "summary": "Update status page", "security": [{ "bearerAuth": [] }] }, "delete": { "summary": "Delete status page", "security": [{ "bearerAuth": [] }] } },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "JWT or API key (pw_live_...)" }
    }
  }
}